third_party/blink/web_tests/external/wpt/largest-contentful-paint/image-upscaling-expected.txt

This is a testharness.js-based test.
PASS Non-scaled image should report the natural size
PASS A downscaled image (width/height) should report the displayed size
PASS A downscaled image (using scale) should report the displayed size
PASS An upscaled image (width/height) should report the natural size
FAIL An upscaled image (using scale) should report the natural size assert_equals: expected 14098 but got 34652
PASS An upscaled image (using object-size) should report the natural size
PASS An intersecting element with a partial-intersecting image (object-position) should report the image intersection
PASS A background image larger than the container should report the container size
PASS A background image smaller than the container should report the natural size
FAIL A scaled-down background image should report the background size assert_equals: expected 100 but got 14098
Harness: the test ran to completion.
